speeeed
before it would get really slow over time when repeatedly loading
sheets. this was because clearing the solve steps was in calc. when
loading a sheet calc isn't ran, so the steps weren't getting cleared and
just growing. this gave mq tons of work (also i think there's an issue
with removing the placeholder characters, if done repeatedly weird stuff
happens). just moved clearing it to the show solve steps function so it
always clears it